requirements.txt
aiogram
fastapi
gspread
google-auth
motor
python-dotenv
uvicorn
